refactor(cli): minor refactor in safe-cli and fix issues reported by clippy in several places
