+++ safe-cli/operations/auth_daemon.rs
@@ -6,21 +6,11 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use dirs;
-use log::debug;
 use prettytable::Table;
 use rpassword;
 use safe_api::{AuthAllowPrompt, Safe, SafeAuthdClient};
-use std::fs::{DirBuilder, File};
-use std::io::{self, Read, Write};
-use std::path::Path;
+use std::io::{self, Write};
 use std::process::Command;
-
-const APP_ID: &str = "net.maidsafe.cli";
-const APP_NAME: &str = "SAFE CLI";
-const APP_VENDOR: &str = "MaidSafe.net Ltd";
-const AUTH_CREDENTIALS_FOLDER: &str = ".safe";
-const AUTH_CREDENTIALS_FILENAME: &str = "credentials";
 
 pub fn authd_run_cmd(command: &str) -> Result<(), String> {
     let authd_bin_path = get_authd_bin_path();
@@ -134,6 +124,7 @@ pub fn authd_subscribe(
     Ok(())
 }
 
+#[allow(dead_code)]
 pub fn authd_subscribe_url(
     safe_authd: &SafeAuthdClient,
     notifs_endpoint: String,
@@ -151,58 +142,6 @@ pub fn authd_unsubscribe(
     println!("Sending request to unsubscribe...");
     safe_authd.unsubscribe(&notifs_endpoint)?;
     println!("Unsubscribed successfully");
-    Ok(())
-}
-
-pub fn authorise_cli(safe: &mut Safe, port: Option<u16>) -> Result<(), String> {
-    println!("Authorising CLI application...");
-    let (mut file, file_path) = get_credentials_file()?;
-    let auth_credentials = safe
-        .auth_app(APP_ID, APP_NAME, APP_VENDOR, port)
-        .map_err(|err| format!("Application authorisation failed: {}", err))?;
-
-    file.write_all(auth_credentials.as_bytes())
-        .map_err(|err| format!("Unable to write credentials in {}: {}", file_path, err))?;
-
-    println!("SAFE CLI app was successfully authorised");
-    println!("Credentials were stored in {}", file_path);
-    Ok(())
-}
-
-pub fn clear_credentials() -> Result<(), String> {
-    let (_file, file_path) =
-        get_credentials_file().map_err(|err| format!("Failed to clear credentials. {}", err))?;
-
-    println!("Credentials were succesfully cleared from {}", file_path);
-    Ok(())
-}
-
-pub fn safe_connect(safe: &mut Safe) -> Result<(), String> {
-    debug!("Connecting...");
-
-    let file_path = credentials_file_path()?;
-    let mut file = File::open(&file_path)
-        .map_err(|_| "You need to authorise the safe CLI first with 'auth' command")?;
-
-    let mut auth_credentials = String::new();
-    file.read_to_string(&mut auth_credentials)
-        .map_err(|_| format!("Unable to read credentials from {}", file_path))?;
-
-    safe.connect(APP_ID, Some(&auth_credentials))
-        .map_err(|err| {
-            format!(
-                "You need to authorise the safe CLI first with 'auth' command: {}",
-                err
-            )
-        })
-}
-
-#[allow(dead_code)]
-pub fn safe_connect_without_auth(safe: &mut Safe) -> Result<(), String> {
-    debug!("Connecting...");
-
-    safe.connect(APP_ID, None)?;
-
     Ok(())
 }
 
@@ -239,28 +178,6 @@ fn prompt_sensitive(arg: Option<String>, msg: &str) -> Result<String, String> {
         rpassword::read_password_from_tty(Some(msg))
             .map_err(|err| format!("Failed reading string from input: {}", err))
     }
-}
-
-fn credentials_file_path() -> Result<String, String> {
-    let home_path =
-        dirs::home_dir().ok_or_else(|| "Couldn't find user's home directory".to_string())?;
-
-    let path = Path::new(&home_path).join(AUTH_CREDENTIALS_FOLDER);
-    if !Path::new(&path).exists() {
-        println!("Creating ~/{} folder", AUTH_CREDENTIALS_FOLDER);
-        DirBuilder::new().recursive(false).create(&path).unwrap();
-    }
-
-    let path = Path::new(&path).join(AUTH_CREDENTIALS_FILENAME);
-    Ok(path.display().to_string())
-}
-
-fn get_credentials_file() -> Result<(File, String), String> {
-    let file_path = credentials_file_path()?;
-    let file = File::create(&file_path)
-        .map_err(|_| format!("Unable to open credentials file at {}", file_path))?;
-
-    Ok((file, file_path))
 }
 
 fn get_authd_bin_path() -> String {
