+++ safe-cli/subcommands/dog.rs
@@ -8,7 +8,7 @@
 
 use super::helpers::{get_from_arg_or_stdin, xorname_to_hex};
 use super::OutputFmt;
-use crate::operations::auth::safe_connect;
+use crate::operations::safe_net::connect;
 use log::debug;
 use prettytable::Table;
 use safe_api::{NrsMapContainerInfo, Safe, SafeData};
@@ -28,8 +28,8 @@ pub fn dog_commander(
     let url = get_from_arg_or_stdin(cmd.location, None)?;
     debug!("Running dog for: {:?}", &url);
 
-    // TODO: switch to safe_connect_without_auth(safe)?;
-    safe_connect(safe)?;
+    // TODO: switch to connect_without_auth(safe)?;
+    connect(safe)?;
     let content = safe.fetch(&url)?;
     match &content {
         SafeData::FilesContainer {
