+++ safe-cli/subcommands/mod.rs
@@ -6,13 +6,13 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-#[cfg(not(any(feature = "fake-auth", feature = "scl-mock")))]
+// #[cfg(not(any(feature = "fake-auth", feature = "scl-mock")))]
 pub mod auth;
 pub mod cat;
 pub mod container;
 pub mod dog;
-#[cfg(any(feature = "fake-auth", feature = "scl-mock"))]
-pub mod fake_auth;
+// #[cfg(any(feature = "fake-auth", feature = "scl-mock"))]
+// pub mod fake_auth;
 pub mod files;
 mod helpers;
 pub mod keys;
@@ -21,10 +21,10 @@ pub mod safe_id;
 pub mod update;
 pub mod wallet;
 
-#[cfg(not(any(feature = "fake-auth", feature = "scl-mock")))]
+// #[cfg(not(any(feature = "fake-auth", feature = "scl-mock")))]
 use auth::AuthSubCommands;
-#[cfg(any(feature = "fake-auth", feature = "scl-mock"))]
-pub use fake_auth::{self as auth, AuthSubCommands};
+// #[cfg(any(feature = "fake-auth", feature = "scl-mock"))]
+// pub use fake_auth::{self as auth, AuthSubCommands};
 
 use structopt::StructOpt;
